+++ src/main/java/ssammudan/cotree/domain/project/project/dto/ProjectCreateRequest.java
@@ -1,4 +1,4 @@
-package ssammudan.cotree.domain.project.dto;
+package ssammudan.cotree.domain.project.project.dto;
 
 import java.time.LocalDate;
 import java.util.Map;
