+++ src/main/java/ssammudan/cotree/domain/project/project/redis/RedisViewCountScheduler.java
@@ -1,4 +1,4 @@
-package ssammudan.cotree.domain.project.redis;
+package ssammudan.cotree.domain.project.project.redis;
 
 import java.util.HashMap;
 import java.util.Map;
