+++ src/main/java/ssammudan/cotree/domain/project/project/service/ProjectServiceImpl.java
@@ -1,4 +1,4 @@
-package ssammudan.cotree.domain.project.service;
+package ssammudan.cotree.domain.project.project.service;
 
 import java.util.List;
 import java.util.Map;
@@ -12,10 +12,11 @@ import org.springframework.web.multipart.MultipartFile;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import ssammudan.cotree.domain.project.dto.ProjectCreateRequest;
-import ssammudan.cotree.domain.project.dto.ProjectCreateResponse;
-import ssammudan.cotree.domain.project.dto.ProjectInfoResponse;
-import ssammudan.cotree.domain.project.dto.ProjectListResponse;
+import ssammudan.cotree.domain.project.common.helper.ProjectHelper;
+import ssammudan.cotree.domain.project.project.dto.ProjectCreateRequest;
+import ssammudan.cotree.domain.project.project.dto.ProjectCreateResponse;
+import ssammudan.cotree.domain.project.project.dto.ProjectInfoResponse;
+import ssammudan.cotree.domain.project.project.dto.ProjectListResponse;
 import ssammudan.cotree.global.error.GlobalException;
 import ssammudan.cotree.global.response.ErrorCode;
 import ssammudan.cotree.global.response.PageResponse;
